Sending items to firebase database through development side
however, it is commented out as it has no real use in the system yet. Remove the comment symbol if you need to insert more items in the database (you can also insert the items manually through firebase itself but i just think it is easier to do it this way (atleast for me) but do not forget to comment it back afterwards as having it with no real purpose might cause some issues later on.